[front][back] create tutor

+++ gcfa-back/src/main/java/com/kraken/gcfa/services/UserService.java
@@ -42,7 +42,7 @@ public class UserService {
         return tutor.getApprentices();
     }
     
-    public List<User> searchUser(Integer roleId, String search) {    	
+    public List<User> searchUser(Long roleId, String search) {
     	return userRepository.searchUser(roleId, search);
     }
